Makefile: simplify detection of python flags
For the last few years, Python 2 is installed as both
python and python2, so the rules can be simplified. As
defined, INSTALLED_PYTHON was used to define flags which
only make sense on UNIX-y systems, since the paths for
Windows will be completely different. So the whole block
can just be removed.

git-svn-id: https://svn.code.sf.net/p/moose/code/moose/branches/zbigniew_py3k@5233 d1858a74-f715-0410-99c9-c66fe9cce3d6